requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.11
SQLAlchemy==2.0.44
Flask-SQLAlchemy==3.0.5
alembic==1.12.0
qrcode[pil]==8.2
gunicorn==21.2.0
